+++ src/kernel_evm/CHANGES.md
@@ -23,12 +23,13 @@
 - Genesis parent hash is 0xff..ff, all parent hashes are now unique. (!10332)
 
 ### Breaking changes
-- L2 blocks' RLP encoding includes all the fields. (!10386)
 
 ### Internal
 
 - Indexes of accounts are now correctly checked, they are no longer duplicated. (!10395)
 - Remove POW related fields from L2 blocks. (!10421)
+- L2 blocks' RLP encoding includes all the fields. (!10386)
+- Block hash is no longer in Tx object, receipt and BIP. (!10520)
 
 ## Version fc06d63568b1be253088ff5d6e422fd80ed3d2c2